requirements/base.txt
asn1crypto==0.24.0
celery==4.0.2
cryptography==2.1.4
Django==2.2
django-redis==4.8.0
django-redis-cache==1.7.1
psycopg2-binary==2.7.6.1
dj-database-url==0.4.2
django-extensions==1.9.8
django-filter==2.1.0
factory-boy==2.10.0
mock==2.0.0
kombu==4.1.0
djangorestframework==3.9.2
django-rest-auth==0.9.3
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
Markdown==2.6.11
gunicorn==19.9.0
coreapi==2.3.3
django-cors-headers==2.2.0
facebook-sdk==2.0.0
django-solo==1.1.3
boto3==1.9.20
django-storages==1.7.1
Pillow==5.3.0
watchtower==0.5.4
redis==2.10.6
django-safedelete==0.5.1

pandas==0.20.1
scikit-learn==0.20.3
pandasql==0.7.3
